Killing Shrimp

+++ Assets/Scripts/Shrimp/ShrimpStats.cs
@@ -10,7 +10,7 @@ public struct ShrimpStats
     public bool gender;  // True = male, False = female
     public float birthTime;  // The time at the point of the shrimp's birth
     public float hunger;  // Value from 0-100
-    public int illnessLevel;  // Value from 0-100
+    public float illnessLevel;  // Value from 0-100
     public int temperament;  // Value from 0-100
 
     public int salineLevel; // Value from 0-100, starting at 50
